requirements.txt
# Core Dependencies
streamlit==1.39.0
python-dotenv==1.0.1

# LangChain Ecosystem - Latest Stable Versions
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-core==0.3.21
langchain-text-splitters==0.3.2

# Vector Store - UPDATED FOR PYTHON 3.11+
faiss-cpu==1.9.0

# OpenAI
openai==1.55.3

# Document Processing
beautifulsoup4==4.12.3
lxml==5.3.0
html5lib==1.1

# Utilities
requests==2.32.3
urllib3==2.2.3

# Additional helpful packages
typing-extensions>=4.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# SerperAPI for web search
google-search-results==2.4.2